backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
httpx
python-multipart
opencv-python-headless
numpy
paddlepaddle
paddleocr
google-generativeai
python-jose[cryptography]
passlib[bcrypt]
